refactor: dispose all resources for restart function
